Use the same injector setting logic for both mixin and middleware
